SFAT-17 - refactored package structure.
