+++ src/main/java/uk/gov/crowncommercial/dts/scale/service/gm/Application.java
@@ -1,4 +1,4 @@
-package uk.gov.crowncommercial.dts.scale.gm;
+package uk.gov.crowncommercial.dts.scale.service.gm;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
