+++ src/main/java/uk/gov/crowncommercial/dts/scale/service/gm/controller/GuidedMatchController.java
@@ -1,4 +1,4 @@
-package uk.gov.crowncommercial.dts.scale.gm.controller;
+package uk.gov.crowncommercial.dts.scale.service.gm.controller;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
-import uk.gov.crowncommercial.dts.scale.gm.model.Journey;
+import uk.gov.crowncommercial.dts.scale.service.gm.model.Journey;
 
 /**
  * Guided Match Controller.
